+++ gifthospitality/admin.py
@@ -102,7 +102,7 @@ class GiftAndHospitalityAdmin(AdminImportExport):
         "entered_date_stamp",
         "entered_by",
     )
-    search_fields = ["id", "group", "entered_by"]
+    search_fields = ["id", "rep", "group", "entered_by"]
 
     list_filter = ("classification_fk__gift_type", "offer", "action_taken")
 
@@ -118,7 +118,7 @@ class GiftAndHospitalityAdmin(AdminImportExport):
             "value",
             "group",
             "rep",
-            # "grade_fk",
+            "grade_fk",
             "offer",
             "company_rep",
             "company",
