+++ gifthospitality/export_csv.py
@@ -15,7 +15,7 @@ def _export_gh_iterator(queryset):
         "Estimated Value Of Offer (£)",
         "DIT Group Offered To/From",
         "DIT Representative Offered To/From",
-        # "Grade",
+        "Grade",
         "Offer",
         "Company Representative Offered To/From",
         "Company Offered To/From",
@@ -36,9 +36,9 @@ def _export_gh_iterator(queryset):
             obj.venue,
             obj.reason,
             obj.value,
-            obj.group,
             obj.rep,
-            # get_fk_value(obj.grade_fk, "gradedescription"),
+            get_fk_value(obj.grade, "gradedescription"),
+            obj.group,
             obj.get_offer_display(),
             obj.company_rep,
             obj.company,
