+++ gifthospitality/filters.py
@@ -53,8 +53,9 @@ class GiftHospitalityFilter(MyFilterSet):
             "category_fk",
             "classification_fk",
             "value",
+            "rep",
+            "grade_fk",
             "group",
-            # "grade_fk",
             "offer",
             "company",
             "action_taken",
