+++ gifthospitality/models.py
@@ -5,7 +5,43 @@ from core.metamodels import (
     IsActiveModel,
 )
 
-from payroll.models import DITGroup, Grade
+from costcentre.models import CostCentre, DepartmentalGroup, Directorate
+
+
+# salaries data
+# define a choice field for this
+class Grade(BaseModel):
+    grade = models.CharField(primary_key=True, max_length=10)
+    gradedescription = models.CharField("Grade Description", max_length=50)
+    order = models.IntegerField
+
+    def __str__(self):
+        return self.grade
+
+    class Meta:
+        verbose_name = "Grade"
+        verbose_name_plural = "Grades"
+
+
+class DITGroup(IsActiveModel):
+    group_display_name = models.CharField(primary_key=True, max_length=50,)
+    group_code = models.ForeignKey(
+        DepartmentalGroup, on_delete=models.PROTECT, null=True, blank=True
+    )
+    directorate_code = models.ForeignKey(
+        Directorate, on_delete=models.PROTECT, null=True, blank=True
+    )
+    cost_centre = models.ForeignKey(
+        CostCentre, on_delete=models.PROTECT, null=True, blank=True
+    )
+
+    def __str__(self):
+        return self.group_display_name
+
+    class Meta:
+        verbose_name = "DIT Group"
+        verbose_name_plural = "DIT Group"
+        ordering = ["cost_centre"]
 
 
 class GiftAndHospitalityClassification(IsActiveModel):
