+++ gifthospitality/tables.py
@@ -15,8 +15,9 @@ class GiftHospitalityTable(FadminTable):
             "venue",
             "reason",
             "value",
+            "rep",
             "group",
-            # "grade_fk",
+            "grade_fk",
             "offer",
             "company_rep",
             "company",
