+++ payroll/admin.py
@@ -9,13 +9,16 @@ from core.admin import (
     AdminImportExport,
 )
 
+from gifthospitality.models import (
+    DITGroup,
+)
+
 from payroll.import_csv import (
     import_GROUP_class,
     import_HR_class,
     import_grade_class,
 )
 from payroll.models import (
-    DITGroup,
     DITPeople,
     Grade,
 )
